+++ src/Contracts/Repository.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace SimpleRepository\Contracts;
+
+/**
+ * @template T
+ */
+interface Repository
+{
+    /**
+     * @return T
+     */
+    public function getDataSource();
+}