+++ src/Repository.php
@@ -6,13 +6,12 @@ use SimpleRepository\Concerns\Safetyable;
 
 /**
  * @template T
+ *
+ * @deprecated When creating a repository class implement \SimpleRepository\Contracts\Repository interface instead.
  */
 abstract class Repository
 {
     use Safetyable;
-
-    /** @var T */
-    protected $dataSource;
 
     /**
      * @return T