+++ CHANGELOG.md
@@ -2,6 +2,57 @@
 
 ## Unreleased
 
+* Remove abstract class Repository
+
+## v4.2.3 - 2024-10-31
+
+* Restore repository interface
+* Update stubs repository model
+
+## v4.2.2 - 2024-10-28
+
+* Remove dataSource property from the repository
+* Fix pass filters to getBuilder
+* Fix check getBuilder method in getPagination
+
+## v4.2.1 - 2024-10-27
+
+* Fix set property in HasEloquentSupport
+
+## v4.2.0 - 2024-10-27
+
+* Add FilterDTO and FilterAdapter
+* Update laravel pint and github action check format code
+
+## v4.1.2 - 2024-10-23
+
+* Update message handleSafely in Repository
+
+## v4.1.1 - 2024-10-23
+
+* Update comment in stubs
+* Add use trait into Repository
+
+## v4.1.0 - 2024-10-23
+
+* Add filter deleted to HasFilter
+
+## v4.0.2 - 2024-10-23
+
+* Remove parent interface Repository
+
+## v4.0.1 - 2024-10-23
+
+* Add HasEloquentSupport
+* Rename from SimpleRepository\Traits to SimpleRepository\Concerns
+* Add stubs/repository.contract.model.stub
+* Remove SimpleRepository\Contracts\Repository
+* Split eloquent support components in repository into a trait
+
+## v3.1.3 - 2024-09-09
+
+* Add generic to Attributes (ModelFactory and ServiceFactory)
+
 ## v3.1.2 - 2024-05-02
 
 * Formatting